requirements.txt
django==4.1.2
psycopg2-binary==2.9.5
python-decouple==3.6
djangorestframework==3.14.0
django-filter==22.1
